Benchmark cases development and ordinal moral theory updates

+++ scripts/AbstractExperiments.py
@@ -11,15 +11,16 @@ import copy
 import time
 import requests
 import atexit
-
+from scripts.TexTables import SaveDataFrameToTexTemplate
 
 
 class ExperimentRunner:
     time_columns = ['Total_time', "Heuristic_time", 'Plan_time', 'Mehr_time', 'CQ1_time', 'CQ2_time', 'Sol_time', 'Out_time', 'Sol_reduce_time']
-    def __init__(self, domain, configs:list, outFolder=None) -> None:
+    def __init__(self, domain, configs:list=None, outFolder=None) -> None:
         self.domain = domain
         self.configs = configs
         self.planner = os.getcwd() + "/MPlan/cmake-build-release-clang/MPlan"
+        
         self.horizon = 3
         self.budget = 18
 
@@ -36,12 +37,18 @@ class ExperimentRunner:
             self.outputFolder = outFolder
         else:
             self.outputFolder = os.getcwd() + f"/Data/Experiments/{domain}/{datetime.now().strftime("%Y-%m-%d %H:%M:%S")}"
+        
+        self.texTablesFolder = os.getcwd() + "/Data/TexTables"
         self.mdpFolder = f"{self.outputFolder}/mdps"
+        self.figuresFolder = f"{self.outputFolder}/Figures"
         self.rawOutFolder = f"{self.outputFolder}/raw"
+        self.texOutFolder = f"{self.outputFolder}/tex"
 
         os.makedirs(self.outputFolder, exist_ok=True)
+        os.makedirs(self.figuresFolder, exist_ok=True)
         os.makedirs(self.mdpFolder, exist_ok=True)
         os.makedirs(self.rawOutFolder, exist_ok=True)
+        os.makedirs(self.texOutFolder, exist_ok=True)
 
     def log(self, msg:str, level:int):
         if level > self.loglevel:
@@ -53,11 +60,6 @@ class ExperimentRunner:
     def makePlanOutFileName(self, confName:str, confRep:int, envRep:int):
         return f"{self.rawOutFolder}/{confName}_con{confRep}_rep{envRep}.json"
 
-    def buildEnvironments(self, configRepetitions:int=1):
-        for i in range(configRepetitions):
-            for conf in self.configs:
-                MDPFactory.buildEnvToFile(self.domain, fileOut=self.makeMdpFileName(conf['Name'], i), **conf)
-                
 
     def run(self, configRepetitions=1, envRepetitions=1):
         # Generate Environments.
@@ -65,22 +67,34 @@ class ExperimentRunner:
         self.buildEnvironments(configRepetitions)
         self.log("*********Built all environments!*********",0)
 
-
         # Call planner on all envs
         self.log("*********Planning...*********",0)
+        self.runPlanner(configRepetitions, envRepetitions)
+        self.log("*********Finished planning!*********",0)
+
+        # Extract all consideration tags from across output files
+        # (inefficient reading each file twice.)
+        self.extractData(configRepetitions, envRepetitions)
+
+
+    def buildEnvironments(self, configRepetitions:int=1):
+        for i in range(configRepetitions):
+            for conf in self.configs:
+                MDPFactory.buildEnvToFile(self.domain, fileOut=self.makeMdpFileName(conf['Name'], i), **conf)
+    
+    def runPlanner(self, configRepetitions=1, envRepetitions=1):
         for conf in self.configs:
             for conf_rep in range(configRepetitions):
                 inFile = self.makeMdpFileName(conf['Name'], conf_rep)
                 for env_rep in range(envRepetitions):
                     self.log(f"Config Name: {conf["Name"]}; Env Repetition: {env_rep}",0)
                     outFile = self.makePlanOutFileName(conf['Name'], conf_rep, env_rep)
-                    result = subprocess.run([self.planner, "--debug", "3", inFile, outFile])
+                    result = subprocess.run([self.planner, "--debug", "0", inFile, outFile])
                     result.check_returncode()
             self.log(f"Finished env on {conf["Name"]}.",0)
-        self.log("*********Finished planning!*********",0)
 
-        # Extract all consideration tags from across output files
-        # (inefficient reading each file twice.)
+
+    def extractData(self, configRepetitions=1, envRepetitions=1):
         self.con_tags = []
         for conf_rep in range(configRepetitions):
             for conf in self.configs:
@@ -102,6 +116,7 @@ class ExperimentRunner:
                     outFile = self.makePlanOutFileName(conf['Name'], conf_rep, env_rep)
                     with open(outFile, 'r') as file:
                         json_data = json.load(file)
+                        bestPolicyIdx = json_data['Solutions_Order'][0]
                         entry = {
                             "Config_name": conf['Name'],
                             "Conf_rep": conf_rep,
@@ -125,13 +140,14 @@ class ExperimentRunner:
                             "Backups": json_data['Backups'],
                             "Iterations": json_data['Iterations'],
                             "Horizon": json_data['Horizon'],
-                            "Min_non_accept": json_data["Solutions"][json_data['Solutions_Order'][0]]["Acceptability"],
+                            "Min_non_accept": json_data["Solutions"][bestPolicyIdx]["Acceptability"],
                             "Num_of_min_non_accept": json_data['Num_Min_Non_Acceptability'],
-                            "Num_of_sols": len(json_data["Solutions"])
+                            "Num_of_sols": len(json_data["Solutions"]),
+                            "Total_Attacks": json_data["Total_Attacks"]
                         }
                         for tag in self.con_tags:
-                            if (tag in json_data["Solutions"][0]["Expectation"].keys()):
-                                entry[tag] = json_data["Solutions"][0]["Expectation"][tag]
+                            if (tag in json_data["Solutions"][bestPolicyIdx]["Expectation"].keys()):
+                                entry[tag] = json_data["Solutions"][bestPolicyIdx]["Expectation"][tag]
                             else:
                                 entry[tag] = "N/A"
                         self.data.append(entry)
@@ -156,7 +172,8 @@ class ExperimentRunner:
         resp = requests.post("http://localhost:18080/MDP", json=req)
         if (resp.status_code!=200):
             self.log(f"MPlan Server Error: {resp.reason}",0)
-            self.ServerProcess.terminate()
+            if not self.ServerProcess is None:
+                self.ServerProcess.terminate()
             return
         dat = resp.json
         resp.close()
@@ -207,9 +224,20 @@ class ExperimentRunner:
         average_durations.to_csv(self.getDurationsFilePath())
         
         # Save durations by theory csv
-        theoryTimes = df.groupby(["Config_name", "Conf_rep"], sort=False)[['Total_time', "Heuristic_time", 'Plan_time', 'Mehr_time', 'Sol_time', 'Out_time', 'Sol_reduce_time', "Expanded_states", "Iterations", "Backups", "Min_non_accept", "Num_of_min_non_accept", "Num_of_sols"]].mean()
-        theoryTimes.round(4)
+        agg_rules = {
+            'Heuristic_time': 'mean',
+            'Plan_time': 'mean',
+            'Sol_time': 'mean',
+            'Mehr_time': 'mean',
+            'Total_time': 'mean',
+            'Backups': 'mean',
+            'Num_of_sols': 'mean'
+        }
+        theoryTimes = df.groupby('Config_name', sort=False).agg(agg_rules).reset_index()
         theoryTimes.to_csv(self.getDurationsByTheoryFilePath())
+
+        print(theoryTimes)
+        SaveDataFrameToTexTemplate(theoryTimes, f"{self.texTablesFolder}/Time_table.tex", f"{self.texOutFolder}/time_table.tex")
 
         # Check consideration worth is the same across same config.
         cols = ["Config_name", "Conf_rep", "Env_rep", "Min_non_accept", "Num_of_min_non_accept"]
@@ -223,13 +251,14 @@ class ExperimentRunner:
         theoryResults = df[cols].groupby('Config_name', sort=False).first()
         theoryResults.to_csv(self.getTheoryExpectationsFilePath())
 
+
     def loadResults(self, of):
         self.outputFolder = of
         df = pd.read_csv(self.getAllDataFilePath())
         self.data = df.to_dict()
 
 
-       #TODO update strings below. 
+    #TODO update strings below. 
     def plotTimeResults(self, df=None):
         if (df==None):
             df = pd.DataFrame(self.data)
@@ -283,4 +312,40 @@ class ExperimentRunner:
         plt.savefig(self.outputFolder + "time_components_percentage_vs_horizon.png")  # Save the plot
         plt.show()
 
-            
+    def plotParetoGraph(self, configName:str, conf_rep:int, con_one_label:str, con_two_label:str, env_rep:int=0, fileName:str="ParetoGraph.png"):
+        outFile = self.makePlanOutFileName(configName, conf_rep, env_rep)
+
+        with open(outFile, 'r') as file:
+            json_data = json.load(file)
+        x = []
+        for sol in json_data["Solutions"]:
+            x.append([float(sol["Expectation"][con_one_label]), float(sol["Expectation"][con_two_label]), sol["Acceptability"]])
+        
+        df = pd.DataFrame(x, columns=[con_one_label, con_two_label, "Non-acceptability"])
+        
+        # FIX: Sort so low non-acceptability (green) is drawn LAST (on top)
+        df = df.sort_values(by="Non-acceptability", ascending=False)
+
+        plt.figure(figsize=(10,7))
+        scatter = plt.scatter(
+            df[con_one_label], 
+            df[con_two_label], 
+            c=df["Non-acceptability"], 
+            cmap='RdYlGn_r', 
+            edgecolors='black',
+            linewidth=0.3, # Thinner line helps reduce clutter
+            alpha=0.8,     # Slight transparency helps see density
+            s=35           # Explicitly set size
+        )
+        plt.xlabel(con_one_label)
+        plt.ylabel(con_two_label)
+        plt.title(f"Expected moral worth of policies under {con_one_label} and {con_two_label}")
+        
+        cbar = plt.colorbar(scatter)
+        cbar.set_label('Non-acceptability (Lower = Green, Higher = Red)')
+
+        plt.grid(True, linestyle='--', alpha=0.6)
+        plt.tight_layout()
+
+        plt.savefig(f"{self.figuresFolder}/{fileName}")
+        plt.show()
